kpatch: Add subcommand '--active-functions' to adjust new sysfs attribute 'stack_order' of livepatch
Add an subcommand of kpatch list with '--active-functions' option to adjust
kernel new attribute 'stack_order' of livepatch of kernel v6.13 or later.

Now, using 'kpatch list --active-functions' can output the enabling function
in the system and the relationship from the enabling function to its
object and its related module.

For older kernel, which is not support 'stack_order' attribute, if there
are just one klp module loaded in the system, we support to output the
active functions. However, if there are more than one klp module loaded,
we can not output the active functions becase the information without
'stack_order' is not accurate.

Suggested-by: Joe Lawrence <joe.lawrence@redhat.com>
Signed-off-by: Wardenjohn <zhangwarden@gmail.com>